+++ R/utils.R
@@ -70,3 +70,19 @@ tz <- function(x) {
     tz <- "UTC"
   tz
 }
+
+check_class_internal <- function(x,
+                         values,
+                         x_name = substitute(x),
+                         error = TRUE) {
+  x_name <- deparse_x_name(x_name)
+
+  check_flag_internal(error)
+
+  class <- class(values)[1]
+
+  if (!inherits(x, class)) {
+    on_fail(x_name, " must be class ", class, error = error)
+  }
+  invisible(x)
+}