+++ NEWS.md
@@ -1,6 +1,7 @@
-- check_names can now check names are unique
+- added check_inherits and check_classes
+- check_names can check names are unique
 - only check unique in check_named when unique = TRUE
-- check_colnames and check_names now accept (col)names = character(0)
+- check_colnames and check_names accept (col)names = character(0)
 
 # checkr 0.0.1